% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mupp_parameters.R
\name{estimate_mupp_params}
\alias{estimate_mupp_params}
\title{Estimate MUPP Parameters}
\usage{
estimate_mupp_params(resp, items, method = "MCMC", ...)
}
\arguments{
\item{resp}{a data.frame of (at least) [person, item, n_dims, resp]}

\item{items}{a data.frame of (at least) [item, statement, dim]}

\item{method}{the estimation method (MCMC is the only one that works now)}

\item{...}{other parameters to add later}
}
\value{
a list of [theta, params] arrays with the third dimension indicating
        the iteration
}
\description{
Estimate MUPP statement and person parameters given item responses and item
properties using MCMC
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
