% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_crossprod_dims}
\alias{find_crossprod_dims}
\title{Find Dimensions given Cross-Product Column}
\usage{
find_crossprod_dims(dim12, n_dims, init = 0L)
}
\arguments{
\item{dim12}{the column of the matrix or data.frame}

\item{n_dims}{the total number of dimensions}

\item{init}{an integer indicating the initial starting value for the set
of integers included in the permutation. See Details.}
}
\value{
A vector/array indicating the cross-product dimensions.
}
\description{
Given the column indexing the appropriate cross-product and the total number
of dimensions, find the two dimensions leading to the cross-product.
}
\details{
\code{init} is useful for indexing C vs R code. If \code{init = 0},
         then the indices will work with 0 indexed languages, such as C or
         Python. If \code{init = 1}, then the indices will work with 1 indexed
         languages, such as R.

         This function assumes that the first n_dims columns are the dims,
         the next n_dims - 1 columns are the cross-products of dimension 1 with
         the remaining dims, the next n_dims - 2 columns are the cross-products
         of dimension 2 with the remaining dims > 2, etc.
}
\seealso{
\code{\link{find_crossprod_column}}
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
