% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-mupp.R
\docType{package}
\name{mupp-package}
\alias{mupp}
\alias{mupp-package}
\title{mupp: Multi-Unidimensional Pairwise Preference (MUPP) IRT Model}
\description{
Probability functions to describe the MUPP model, simulate responses
 to the MUPP model, and estimate parameters of the MUPP model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://swnydick.github.io/mupp/index.html}
  \item \url{https://github.com/swnydick/mupp}
}

}
\author{
\strong{Maintainer}: Steven Nydick \email{swnydick@gmail.com} (\href{https://orcid.org/0000-0002-2908-1188}{ORCID})

}
\keyword{internal}
