% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mupp.R
\docType{package}
\name{mupp}
\alias{mupp}
\alias{mupp-package}
\title{mupp}
\description{
Probability functions to describe the MUPP model, simulate responses to the
MUPP model, and estimate parameters of the MUPP model.
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
