% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permutation}
\alias{permutation}
\alias{find_all_permutations}
\alias{find_permutation_order}
\alias{find_permutation_index}
\title{Find All Permutations of Consecutive Integers}
\usage{
find_all_permutations(n, init = 0L)

find_permutation_order(n, index = 1L, init = 0L)

find_permutation_index(order)
}
\arguments{
\item{n}{an integer greater than 0}

\item{init}{an integer indicating the initial starting value for the set
of integers included in the permutation. See Details.}

\item{index}{an integer between 1 and n indication which permutation to
select (using R indexing rather than C indexing)}

\item{order}{an integer vector indicating the response/permutation order
that should be checked against}
}
\value{
A matrix of size n! x n, where each row is a unique permutation
}
\description{
Given an consecutive integer vector of length n, find all permutations of
that vector.
}
\details{
\code{init} is useful for indexing C vs R code. If \code{init = 0},
then the indices will work with 0 indexed languages, such as C or
Python. If \code{init = 1}, then the indices will work with 1 indexed
languages, such as R.
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
