% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_ggum.R
\name{prior_ggum}
\alias{prior_ggum}
\alias{r_thetas_prior}
\alias{r_alpha_prior}
\alias{r_delta_prior}
\alias{r_tau_prior}
\title{Prior GGUM Distributions}
\usage{
r_thetas_prior(n, ...)

r_alpha_prior(n, ...)

r_delta_prior(n, ...)

r_tau_prior(n, ...)
}
\arguments{
\item{n}{number of observations required for the simulation}
}
\description{
Simulate prior distribution parameters for the GGUM model
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
