% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mupp.R
\name{simulate_mupp_params}
\alias{simulate_mupp_params}
\title{Simulate MUPP Parameters}
\usage{
simulate_mupp_params(
  n_persons = 1,
  n_items = 1,
  n_dims = 2,
  max_item_dims = NULL,
  unidim_items = FALSE
)
}
\arguments{
\item{n_persons}{integer indicating the number of persons}

\item{n_items}{integer indicating the number of items}

\item{n_dims}{integer > 2 indicating the total number of dimensions}

\item{max_item_dims}{integer indicating the maximum dimensions on any item}

\item{unidim_items}{boolian indicating whether an item can load on only a
single dimension or must load on multiple dimensions}
}
\value{
a list with items/persons that conform to the MUPP model, as expected
in package functions
}
\description{
Generate parameters/thetas that conform to the MUPP model.
}
\details{
For the purposes of parameter generation, each item will be assumed
to have between 2 and n_dims dimensions, where the number of dimensions
for an item is randomly generated from a discrete uniform distribution.
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
