% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mupp_parameters_.R
\name{estimate_mupp_params}
\alias{estimate_mupp_params}
\title{Estimate MUPP Parameters}
\usage{
estimate_mupp_params(resp, items, method = "MCMC", control = list(), ...)
}
\arguments{
\item{resp}{a data.frame of (at least) [person, item, resp]}

\item{items}{a data.frame of (at least) [item, statement, dim]}

\item{method}{the estimation method (MCMC is the only one that works now)}

\item{control}{a list of parameters to control the algorithm. See details.}

\item{...}{other parameters to pass to the estimation algorithm. See details.}
}
\value{
a list of [theta, params] arrays with the third dimension indicating
        the iteration as well as means and sds
}
\description{
Estimate MUPP statement and person parameters given item responses and item
properties using MCMC
}
\details{
\itemize{
  \item{For MCMC, additional parameters include
    \describe{
      \item{initial_params}{a named list with names indicating the parameter
                            and values indicating the starting values/initial
                            parameter estimates}
      \item{fixed_params}{a character vector with elements that are fixed
                          to their initial parameters for the entire estimation
                          algorithm}
    }
  }
}

\itemize{
  \item{For MCMC, control parameters include

    \describe{
      \item{n_iters}{total number of iterations.}
      \item{n_burnin}{number of iterations to throw away when calculating
                      summary statistics.}
      \item{step_size_sd}{the standard deviation of the step size for subsequent
                          Metropolis-Hastings draws.}
    }
  }
}
}
\examples{
\dontrun{
set.seed(3452345)

# simulate parameters and responses to the model
# (assumption is that params/resp will follow conventions)
params <- simulate_mupp_params(n_persons     = 100,
                               n_items       = 300,
                               n_dims        = 9,
                               max_item_dims = 2,
                               unidim_items  = TRUE)
resp   <- do.call(simulate_mupp_resp,
                  params)

# thetas for comparison
thetas <- tidyr::spread(params$persons,
                        key   = "dim",
                        value = "theta")[ , -1]
items  <- params$items

# estimating thetas using algorithm (one start for comparison purposes)
est_params  <- estimate_mupp_params(resp    = resp$resp,
                                    items   = resp$items,
                                    method  = "MCMC",
                                    control = list(n_iters  = 1000,
                                                   n_burnin = 500),
                                    initial_params = list(delta = sign(items$delta)))

# correlating (not great, but small iters and few people)
diag(cor(thetas, est_params$means$thetas))
cor(items$alpha, est_params$means$alpha)
cor(items$delta, est_params$means$delta)
cor(items$tau,   est_params$mean$tau)
}

}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
