% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mupp_thetas.R
\name{estimate_mupp_thetas}
\alias{estimate_mupp_thetas}
\title{Estimate MUPP Thetas}
\usage{
estimate_mupp_thetas(resp, items, method = c("newton_r", "bfgs_r", "newton_c",
  "bfgs_c", "MCMC"), control = list(), ...)
}
\arguments{
\item{resp}{a data.frame of (at least) [person, item, resp]}

\item{items}{a data.frame of (at least) [item, statement, dim, alpha, delta, tau]}

\item{method}{the estimation method ("newton_r", "bfgs_r", "newton_c", "bfgs_c",
"MCMC")}

\item{...}{other parameters to pass to the method}
}
\value{
a list of [thetas, sems, hessians]
}
\description{
Estimate MUPP person parameters given item responses, item parameters, and
item properties.
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}
}
